pyproject.toml
[tool.poetry]
name = "linien-influxdb"
version = "2.0.0.dev0"
description = "Log status of linien locking application to influxdb"
authors = ["Bastian Leykauf <leykauf@physik.hu-berlin.de>"]
license = "GNU General Public License v3 (GPLv3)"
readme = "README.md"
packages = [{include = "linien_influxdb"}]
homepage = "https://pypi.org/project/linien-influxdb"
repository = "https://github.com/linien-org/linien/"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent"
]

[tool.poetry.dependencies]
python = "^3.7"
influxdb-client = "^1.27.0"
configparser = "^5.2.0"
click = "^8.1.2"
linien-python-client = "^0.5.3"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
pre-commit = "^2.20.0"
flake8 = "^5.0.4"
rope = "^1.3.0"
isort = "^5.10.1"

[tool.poetry.scripts]
influxdb-logger = "influxdb_logger.app:main"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
